# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Oficina Mecânica Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ordens:
    get:
      operationId: listOrdens
      tags: [ordens]
      summary: List all service orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [orcamento, em_andamento, finalizado, entregue]
      responses:
        "200":
          description: List of service orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrdemServico"
    post:
      operationId: createOrdem
      tags: [ordens]
      summary: Create a service order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrdemServico"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServico"
  /ordens/{id}:
    get:
      operationId: getOrdem
      tags: [ordens]
      summary: Get a service order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServicoDetalhada"
    put:
      operationId: updateOrdem
      tags: [ordens]
      summary: Update a service order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrdemServico"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServico"
    delete:
      operationId: deleteOrdem
      tags: [ordens]
      summary: Delete a service order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ordens/{id}/status:
    patch:
      operationId: updateOrdemStatus
      tags: [ordens]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatus"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServico"

  # FINANCEIRO
...
components:
  schemas:
...
    OrdemServico:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteId:
          type: integer
        clienteNome:
          type: string
        veiculoId:
          type: integer
        veiculoPlaca:
          type: string
        veiculoModelo:
          type: string
        responsavel:
          type: string
        status:
          type: string
          enum: [orcamento, em_andamento, finalizado, entregue]
        dataEntrada:
          type: string
        dataPrevisao:
          type: string
        dataFinalizacao:
          type: string
        observacoes:
          type: string
        totalServicos:
          type: number
        totalPecas:
          type: number
        total:
          type: number
        atrasado:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - numero
        - clienteId
        - clienteNome
        - veiculoId
        - veiculoPlaca
        - veiculoModelo
        - responsavel
        - status
        - dataEntrada
        - totalServicos
        - totalPecas
        - total
        - atrasado
        - createdAt
    OrdemServicoDetalhada:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteId:
          type: integer
        clienteNome:
          type: string
        veiculoId:
          type: integer
        veiculoPlaca:
          type: string
        veiculoModelo:
          type: string
        veiculoMarca:
          type: string
        veiculoAno:
          type: integer
        veiculoKm:
          type: integer
        responsavel:
          type: string
        status:
          type: string
          enum: [orcamento, em_andamento, finalizado, entregue]
        dataEntrada:
          type: string
        dataPrevisao:
          type: string
        dataFinalizacao:
          type: string
        observacoes:
          type: string
        checklistEntrada:
          type: object
          additionalProperties: true
        checklistEntrega:
          type: object
          additionalProperties: true
        servicos:
          type: array
          items:
            $ref: "#/components/schemas/OrdemServicoItem"
        pecas:
          type: array
          items:
            $ref: "#/components/schemas/OrdemPecaItem"
        totalServicos:
          type: number
        totalPecas:
          type: number
        total:
          type: number
        atrasado:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - numero
        - clienteId
        - clienteNome
        - veiculoId
        - veiculoPlaca
        - veiculoModelo
        - veiculoMarca
        - veiculoAno
        - responsavel
        - status
        - dataEntrada
        - servicos
        - pecas
        - totalServicos
        - totalPecas
        - total
        - atrasado
        - createdAt
    OrdemServicoItem:
      type: object
      properties:
        id:
          type: integer
        servicoId:
          type: integer
        nome:
          type: string
        valor:
          type: number
      required:
        - id
        - servicoId
        - nome
        - valor
    OrdemPecaItem:
      type: object
      properties:
        id:
          type: integer
        pecaId:
          type: integer
        nome:
          type: string
        codigo:
          type: string
        quantidade:
          type: integer
        valorUnitario:
          type: number
        valorTotal:
          type: number
      required:
        - id
        - pecaId
        - nome
        - quantidade
        - valorUnitario
        - valorTotal
    CreateOrdemServico:
      type: object
      properties:
        clienteId:
          type: integer
        veiculoId:
          type: integer
        responsavel:
          type: string
        status:
          type: string
          enum: [orcamento, em_andamento, finalizado, entregue]
        dataEntrada:
          type: string
        dataPrevisao:
          type: string
        observacoes:
          type: string
        checklistEntrada:
          type: object
          additionalProperties: true
        servicos:
          type: array
          items:
            $ref: "#/components/schemas/AddServicoItem"
        pecas:
          type: array
          items:
            $ref: "#/components/schemas/AddPecaItem"
      required:
        - clienteId
        - veiculoId
        - responsavel
        - dataEntrada
    UpdateOrdemServico:
      type: object
      properties:
        clienteId:
          type: integer
        veiculoId:
          type: integer
        responsavel:
          type: string
        status:
          type: string
          enum: [orcamento, em_andamento, finalizado, entregue]
        dataEntrada:
          type: string
        dataPrevisao:
          type: string
        dataFinalizacao:
          type: string
        observacoes:
          type: string
        checklistEntrada:
          type: object
          additionalProperties: true
        checklistEntrega:
          type: object
          additionalProperties: true
        servicos:
          type: array
          items:
            $ref: "#/components/schemas/AddServicoItem"
        pecas:
          type: array
          items:
            $ref: "#/components/schemas/AddPecaItem"
    AddServicoItem:
      type: object
      properties:
        servicoId:
          type: integer
        valor:
          type: number
      required:
        - servicoId
        - valor
    AddPecaItem:
      type: object
      properties:
        pecaId:
          type: integer
        quantidade:
          type: integer
        valorUnitario:
          type: number
      required:
        - pecaId
        - quantidade
        - valorUnitario
    UpdateStatus:
      type: object
      properties:
        status:
          type: string
          enum: [orcamento, em_andamento, finalizado, entregue]
        checklistEntrega:
          type: object
          additionalProperties: true
      required:
        - status
...