# lib/api-zod/src/generated/types/ordemServicoDetalhada.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Oficina Mecânica Management System
 * OpenAPI spec version: 0.1.0
 */
import type { OrdemPecaItem } from "./ordemPecaItem";
import type { OrdemServicoDetalhadaChecklistEntrada } from "./ordemServicoDetalhadaChecklistEntrada";
import type { OrdemServicoDetalhadaChecklistEntrega } from "./ordemServicoDetalhadaChecklistEntrega";
import type { OrdemServicoDetalhadaStatus } from "./ordemServicoDetalhadaStatus";
import type { OrdemServicoItem } from "./ordemServicoItem";

export interface OrdemServicoDetalhada {
  id: number;
  numero: string;
  clienteId: number;
  clienteNome: string;
  veiculoId: number;
  veiculoPlaca: string;
  veiculoModelo: string;
  veiculoMarca: string;
  veiculoAno: number;
  veiculoKm?: number;
  responsavel: string;
  status: OrdemServicoDetalhadaStatus;
  dataEntrada: string;
  dataPrevisao?: string;
  dataFinalizacao?: string;
  observacoes?: string;
  checklistEntrada?: OrdemServicoDetalhadaChecklistEntrada;
  checklistEntrega?: OrdemServicoDetalhadaChecklistEntrega;
  servicos: OrdemServicoItem[];
  pecas: OrdemPecaItem[];
  totalServicos: number;
  totalPecas: number;
  total: number;
  atrasado: boolean;
  createdAt: string;
}